clk_pll_16fft.c: 6: yes: MISRA.ETYPE.COMP.ASSIGN.2012
Fix composite expression assignment to wider essential type violations

- Rule 10.6, composite expression assigned to wider essential type
- Fixed 2 violations: explicit cast of clkod to uint32_t in multiplication
- IDs 3642, 3643 resolved, enforced violations completed

+++ plat/ti/k3low/common/drivers/pm/clock/clk_pll_16fft.c
@@ -868,7 +868,7 @@ static uint32_t clk_pll_16fft_internal_set_freq_from_pll_table(struct clk *pll_c
 		 * Note: We break up the calculation in order to avoid a div64.
 		 */
 		clkod_plld = (uint32_t)soc_pll_table[data_pll->pll_entries[i]].plld *
-			     soc_pll_table[data_pll->pll_entries[i]].clkod;
+			     (uint32_t)soc_pll_table[data_pll->pll_entries[i]].clkod;
 
 			actual64 = ((uint64_t) input / clkod_plld) *
 				soc_pll_table[data_pll->pll_entries[i]].pllm;
